Move `ImageMounter`/`ImageMapper` into a separate module.
