+++ examples/dbt_adapter/dbt/adapters/splitgraph/connections.py
@@ -9,7 +9,7 @@ from psycopg2.pool import AbstractConnectionPool
 from splitgraph.core.repository import Repository
 from splitgraph.engine import switch_engine
 from splitgraph.engine.postgres.engine import PostgresEngine
-from splitgraph.splitfile.execution import ImageMapper
+from splitgraph.core.image_mounting import ImageMapper
 
 
 @dataclass
